Latest fight with flexbox via iron-flex.
I altered the profile view to use the official Layout Classes offered by iron-flex instead of using the iron-flex mixins in CSS or flexbox directly. This appears to be the “easy” way according to the Polymer team, but I’m still trying to wrap my head around the million ways to do this!

https://github.com/PolymerElements/iron-flex-layout